+++ aether/transport/socket_packet_queue_manager.h
@@ -26,7 +26,7 @@
 #include "aether/events/multi_subscription.h"
 
 #include "aether/transport/actions/packet_send_action.h"
-#include "aether/transport/low_level/socket_packet_send_action.h"
+#include "aether/transport/socket_packet_send_action.h"
 
 namespace ae {
 template <typename TSocketPacketSendAction,
